+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## master / unreleased
+
+## 1.15.0 in progress
+
 * [CHANGE] Storage: Make Max exemplars config per tenant instead of global configuration. #5016
 * [CHANGE] Alertmanager: Local file disclosure vulnerability in OpsGenie configuration has been fixed. #5045
 * [CHANGE] Rename oltp_endpoint to otlp_endpoint to match opentelemetry spec and lib name. #5067